feat: add transaction validation task
