+++ Cargo.toml
@@ -114,7 +114,7 @@ tokio = { version = "1.21", default-features = false }
 tokio-util = { version = "0.7.4", features = ["codec"] }
 
 ## async
-async-trait = "0.1.58"
+async-trait = "0.1.68"
 futures = "0.3.26"
 pin-project = "1.0.12"
 futures-util = "0.3.25"
