+++ bin/reth/src/node/mod.rs
@@ -206,7 +206,12 @@ impl Command {
         let blockchain_db = BlockchainProvider::new(factory, blockchain_tree.clone())?;
 
         let transaction_pool = reth_transaction_pool::Pool::eth_pool(
-            EthTransactionValidator::new(blockchain_db.clone(), Arc::clone(&self.chain)),
+            EthTransactionValidator::new(
+                blockchain_db.clone(),
+                Arc::clone(&self.chain),
+                ctx.task_executor.clone(),
+                1,
+            ),
             Default::default(),
         );
         info!(target: "reth::cli", "Transaction pool initialized");
@@ -218,14 +223,11 @@ impl Command {
             let client = blockchain_db.clone();
             ctx.task_executor.spawn_critical(
                 "txpool maintenance task",
-                Box::pin(async move {
-                    reth_transaction_pool::maintain::maintain_transaction_pool(
-                        client,
-                        pool,
-                        chain_events,
-                    )
-                    .await
-                }),
+                reth_transaction_pool::maintain::maintain_transaction_pool_future(
+                    client,
+                    pool,
+                    chain_events,
+                ),
             );
             debug!(target: "reth::cli", "Spawned txpool maintenance task");
         }
